# [CHIA-2052] Compressed blocks