--- a/chia/full_node/full_node_api.py
+++ b/chia/full_node/full_node_api.py
@@ -860,9 +860,15 @@ class FullNodeAPI:
                     while not curr_l_tb.is_transaction_block:
                         curr_l_tb = self.full_node.blockchain.block_record(curr_l_tb.prev_hash)
                     try:
-                        new_block_gen = await self.full_node.mempool_manager.create_block_generator(
-                            curr_l_tb.header_hash
-                        )
+                        # TODO: once we're confident in the new block creation,
+                        # switch to it by default
+                        if self.full_node.config.get("original_block_creation", True):
+                            create_block = self.full_node.mempool_manager.create_block_generator
+                        else:
+                            create_block = self.full_node.mempool_manager.create_block_generator2
+
+                        new_block_gen = await create_block(curr_l_tb.header_hash)
+
                     except Exception as e:
                         self.log.error(f"Traceback: {traceback.format_exc()}")
                         self.full_node.log.error(f"Error making spend bundle {e} peak: {peak}")