### Purpose:

Farm blocks with as many transactions as possible.

### Current Behavior:

We first fill the block with transactions, then we compress it. We don't attempt to fit more transactions after compression, leaving blocks under utilized.

### New Behavior:

This patch doesn't change the default behavior, it adds a (temporary) configuration option to opt-in to a new block creation algorithm.

The new block creation function (`create_block_generator2()`) uses the `BlockBuilder` class from `chia_rs`. This class allows building and compressing a block incrementally. So we start with a batch of transactions that add up to the block cost limit and serialize and compress that. Then we keep building batches to add to the block and serialize and compress those until we reach the cost limit, or run out of time.

### Testing Notes:

The new algorithm is slower, because it serializes more transactions into the block, but also because it needs to checkpoint the serializer state in case it exceeds the limit and has to revert.

Here's how it measures up against the existing block creation (which is also bottle-necked by serialization and compression):

(MacBook M1)
```
Profiling create_block_generator()
  output written to: mempool-create-st.png
  time: 8.3498s
  per call: 834.98ms

Profiling create_block_generator2()
  output written to: mempool-create2-st.png
  time: 17.4403s
  per call: 1744.03ms
```

(RPi-5)
```
Profiling create_block_generator()
  output written to: mempool-create-st.png
  time: 46.4750s
  per call: 4647.50ms

Profiling create_block_generator2()
  output written to: mempool-create2-st.png
  time: 50.0050s
  per call: 5000.50ms
```

This is one reason this is not enabled by default yet. There are optimizations of the compression/serialization logic in the works, [here](https://github.com/Chia-Network/clvm_rs/pull/562).

This will significantly speed up compression (see benchmarks in that PR).

Including this new algorithm early gives us a chance to use it on testnet for longer, and maybe even farmers in the wild, before we switch over to it.

## manual tests

On rapsberry PI 5, running a testnet11 node I used the new RPC to create a block generator with the following log output:

```
full_node chia.full_node.full_node: INFO     Beginning simulated block construction from mempool
full_node chia.full_node.mempool  : INFO     Starting to make block, max cost: 10999975980
full_node chia.full_node.mempool  : INFO     adding TX batch, additions: 9 removals: 8 cost: 29549118 total cost: 10914077138
full_node chia.full_node.mempool  : INFO     adding TX batch, additions: 1 removals: 1 cost: 3039460 total cost: 10954508598
full_node chia.full_node.mempool  : INFO     adding TX batch, additions: 1 removals: 1 cost: 3039460 total cost: 10994916058
full_node chia.full_node.mempool  : INFO     create_block_generator2() took 0.0154 seconds. block cost: 10994916058 spends: 10 additions: 11
full_node chia.full_node.full_node: INFO     Simulated block constructed in 0.02 seconds
```

## future performance

Running the mempool benchmark on RPi-5, with the CLVM optimization patch applied, give these results:

```
Profiling create_block_generator()
  output written to: mempool-create-st.png
  time: 49.7158s
  per call: 4971.58ms

Profiling create_block_generator2()
  output written to: mempool-create2-st.png
  time: 21.0602s
  per call: 2106.02ms
```